requirements.txt
streamlit
langchain
python-dotenv
seaborn
pandas
langchain-google-genai